refactor: move PhotoBulkAdd from features/ai to features/bins
